Rename TypedEndpointResponse<> to EndpointResponse<T> and EndpointResponse<T1,T2>
