+++ src/Atc.Rest.Client/EndpointResponse.cs
@@ -2,6 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Net;
 
+#pragma warning disable SA1402 // File may only contain a single type
+
 namespace Atc.Rest.Client
 {
     public class EndpointResponse : IEndpointResponse
@@ -46,5 +48,52 @@ namespace Atc.Rest.Client
             return ContentObject as TResult ??
                    throw new InvalidCastException($"ContentObject is not of type {typeof(TResult).Name}");
         }
+    }
+
+    public class EndpointResponse<T>
+        : EndpointResponse
+        where T : class
+    {
+        public EndpointResponse(EndpointResponse response)
+            : base(response)
+        {
+        }
+
+        public EndpointResponse(
+            bool isSuccess,
+            HttpStatusCode statusCode,
+            string content,
+            T? contentObject,
+            IReadOnlyDictionary<string, IEnumerable<string>> headers)
+            : base(isSuccess, statusCode, content, contentObject, headers)
+        {
+        }
+
+        public T? SuccessContent => IsSuccess ? CastContent<T>() : null;
+    }
+
+    public class EndpointResponse<TSuccessContent, TFailureContent>
+        : EndpointResponse
+        where TSuccessContent : class
+        where TFailureContent : class
+    {
+        public EndpointResponse(EndpointResponse response)
+            : base(response)
+        {
+        }
+
+        public EndpointResponse(
+            bool isSuccess,
+            HttpStatusCode statusCode,
+            string content,
+            object? contentObject,
+            IReadOnlyDictionary<string, IEnumerable<string>> headers)
+            : base(isSuccess, statusCode, content, contentObject, headers)
+        {
+        }
+
+        public TSuccessContent? SuccessContent => IsSuccess ? CastContent<TSuccessContent>() : null;
+
+        public TFailureContent? FailureContent => !IsSuccess ? CastContent<TFailureContent>() : null;
     }
 }
